moves header and footer to subfolders
