+++ src/components/Layout/Header/Header.css
@@ -1,0 +1,33 @@
+header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100vw;
+  padding-top: 1rem;
+  margin: 0;
+}
+
+header a {
+  color: var(--pennyroyal);
+  font-weight: 400;
+  font-size: 1.5rem;
+  margin: 1rem;
+  padding: 0;
+  text-decoration: none;
+  cursor: pointer;
+  font-family: 'Barlow', sans-serif;
+  border-bottom: solid transparent 1px;
+}
+
+header a:hover,
+header a:focus {
+  color: var(--aqua);
+  border-bottom: solid var(--white-ish) 1px;
+}
+
+.title {
+  color: white;
+  font-size: 2rem;
+  font-family: 'Major Mono Display', monospace;
+  text-decoration: none;
+}
