+++ src/components/Layout/Layout.jsx
@@ -1,12 +1,12 @@
-import Header from './Header';
-import Footer from './Footer';
+import Header from './Header/Header.jsx';
+import Footer from './Footer/Footer.jsx';
 import styles from './Layout.css';
 
 export default function Layout({ children }) {
   return (
     <div className={styles.layout}>
       <Header />
-        <main >{children}</main>
+      <main>{children}</main>
       <Footer />
     </div>
   );
